+++ src/index.ts
@@ -1,17 +1,20 @@
 import { AuthService } from './services/auth';
 import { MarkkoConfig } from './types';
-import { VendorsAPI } from './api/vendors';
+import { AddressesAPI } from './api/addresses';
 import { BlogsAPI } from './api/blogs';
+import { VendorsAPI } from './api/vendors';
 
 export class MarkkoSDK {
   private authService: AuthService;
-  public vendors: VendorsAPI;
+  public addresses: AddressesAPI;
   public blogs: BlogsAPI;
+  public vendors: VendorsAPI;
 
   constructor(config: MarkkoConfig) {
     this.authService = new AuthService(config);
-    this.vendors = new VendorsAPI(config, this.authService);
+    this.addresses = new AddressesAPI(config, this.authService);
     this.blogs = new BlogsAPI(config, this.authService);
+    this.vendors = new VendorsAPI(config, this.authService);
   }
 
   async getAccessToken(): Promise<string> {